pyproject.toml
[tool.poetry]
name = "fastapi_react_typescript_template"
version = "0.1.0"
description = ""
authors = ["BuRny <gamingburny@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7 <3.11"
loguru = "^0.5"
fastapi = "^0.68"
uvicorn = {extras = ["standard"], version = "^0.15"}
# Enable connection to postgres
psycopg2-binary = "^2.9"
# Enable connection to mongodb
pymongo = "^3.12"

[tool.poetry.dev-dependencies]
# Autoformatter
yapf = "^0.31"
# Linting
pylint = "^2.8"
# Test library
pytest = "^6.2"
pytest-asyncio = "^0.15"
pytest-cov = "^2.12"
pytest-benchmark = "^3.4"
# Type checking
mypy = "^0.910"
# Requirement for pytest < 3.8
importlib_metadata = "^1.7"
# Statistical testing
hypothesis = "^6.21"
# Cyclomatic complexity checker
radon = "^5.0"
# Enable pre commit hooks
pre-commit = "^2.15"
# e2e tests
seleniumbase = "^1.65"
# Monitor processes
psutil = "^5.8"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.masonry.api"

[tool.mypy]
python_version = "3.7"
ignore_missing_imports = true

[tool.pycln]
all = true

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pylint.design]
# Maximum number of locals for function / method body
max-locals = 25
[tool.pylint.messages_control]
max-line-length = 120
# C0301,E501 2x Line too long
# C0114 module Docstring
# C0115 missing class docstring
# C0116 missing function docstring
# R0913 function with too many arguments
# C0413 import order
# C0411 import order of external libs
# W0511 TODO
# W0105 string statement with no effect
# R0801 duplicate code
# W0621 redefining name from outer score
# C0103 variable name does not conform snake case naming style
# R0903: Too few public methods of a class
# E1101: Class 'SqlMetaclass' has no '__annotations__' member (no-member)
disable = "C0301,E501, C0114, C0115, C0116, R0913, C0413, C0411, W0511, W0105, R0801, W0621, C0103, R0903, E1101"

[tool.yapf]
based_on_style = "pep8"
column_limit = 120
split_arguments_when_comma_terminated = true
dedent_closing_brackets = true